Move app bar related code into app_bar_actions.dart
